pyproject.toml
[project]
name = "rag-jfk"
version = "0.1.0"
description = "Retrivial Augmented Generation With JFK Speeches"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "beautifulsoup4>=4.12.3",
    "google-cloud-storage>=2.19.0",
    "jq>=1.8.0",
    "langchain>=0.3.14",
    "langchain-community>=0.3.14",
    "langchain-core>=0.3.30",
    "langchain-openai>=0.3.1",
    "langchain-pinecone>=0.2.2",
    "langchain_groq>=0.3.2",
    "ruff>=0.11.7",
    "streamlit>=1.44.1",
    "langchain_nvidia_ai_endpoints>=0.3.10"

]
